# PERF: Indexing with pyarrow timestamp & duration dtypes