- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.


pyarrow timestamps:

```
import pandas as pd

N = 100_000

idx = pd.Index(range(N), dtype="timestamp[s][pyarrow]")
idx2 = idx[::2]

%timeit idx.get_indexer_for(idx2)

# 264 ms ± 13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)      -> main
# 1.1 ms ± 168 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  -> PR
```

pyarrow durations:

```
import pandas as pd

N = 100_000

idx = pd.Index(range(N), dtype="duration[s][pyarrow]")
idx2 = idx[::2]

%timeit idx.get_indexer_for(idx2)

# 705 ms ± 21.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)        -> main
# 1.35 ms ± 83.9 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  -> PR
```